move cron job syncCognitoToDb to tasks.subservice
